services/classifier/requirements.txt
karton-core
pefile
